moved "users-model" to auth folder and changed name to "auth-model"
